Add support for `camt.054.001.08`
V08 introduces substructure for related parties (`Party40Choice` in the
spec). They can now be either a private party (like before) or a
financial institution (new).

Because our existing model is too simple, we will read the new financial
institution name and postal address into a `RelatedPartyTypeInterface`,
but we will not read anything else, and we will not be able to know if
it is a private party or a financial institution.

Similarly, an `Entry` status code now support proprietary code. It will
be read, but we will not be able to know if it is a standard one (eg:
`"BOOK"`), or a proprietary one.

Fixes #126