`Amt` and `TtlChrgsAndTaxAmt` may be 0
When parsing an XML-File which contains an amount of a ChargesRecord that
has "0" as its value, the amount was not being set on the ChargesRecord object.

Example:

```xml
<Chrgs>
    <Rcrd>
        <Amt Ccy="CHF">0</Amt>
        ...
    </Rcd>
</Crgs>
```

Now, if the amount is available in XML, then it will be set on the
ChargesRecord object, even with a value of 0.

Correct